# ai-version/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API (AI Generated)",
    "version": "1.0.0",
    "description": "REST API for managing tasks generated by AI specification."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root Info",
        "responses": {
          "200": { "description": "Successful operation" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": { "description": "Server is healthy" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": { "description": "List of tasks" }
        }
      },
      "post": {
        "summary": "Create task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskInput"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Not Found" }
        }
      },
      "put": {
        "summary": "Update task",
        "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskInput"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" }
        }
      },
      "delete": {
        "summary": "Delete task",
        "parameters": [{ "$ref": "#/components/parameters/TaskId" }],
        "responses": {
          "204": { "description": "Deleted" },
          "404": { "description": "Not Found" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "TaskId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      }
    },
    "schemas": {
      "CreateTaskInput": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "Buy grocery" }
        }
      },
      "UpdateTaskInput": {
        "type": "object",
        "properties": {
          "title": { "type": "string", "example": "Buy grocery updated" },
          "done": { "type": "boolean", "example": true }
        }
      }
    }
  }
}